ml/requirements.txt
psycopg2-binary>=2.9.9
scikit-learn>=1.4.0
pandas>=2.1.0
numpy>=1.26.0
joblib>=1.3.0
boto3>=1.34.0
